Update overview.md
I removed Veracity Adapter for Power BI, as the user are not familiar with it but more with the service name.
Icons to be shown depend on settings on the report. Therefor added may see instead of will see
You used File icon, I changed to document icon. The icon is only shown if a service have added documents to be available.