Update overview.md
'Read a report' is not what I think we should say. View and interact is better.
I changed it here and also added the new Export to PDF & PPT in the next release. 

Please double check the the use of Read other places we should change to view and interact. 
Please convince me if Read is better to use.